+++ modules/client/src/addresslistVoting/internal/graphql-queries/members.ts
@@ -1,8 +1,8 @@
 import { gql } from "graphql-request";
 
 export const QueryAddresslistVotingMembers = gql`
-query AddresslistVotingMembers($address: ID!) {
-    addresslistVotingPlugin(id: $address){
+query AddresslistVotingMembers($address: ID!, $block: Block_height) {
+    addresslistVotingPlugin(id: $address, block: $block){
         members {
             address
         }