+++ modules/client/src/multisig/internal/graphql-queries/members.ts
@@ -1,8 +1,8 @@
 import { gql } from "graphql-request";
 
 export const QueryMultisigMembers = gql`
-query MultisigMembers($address: ID!) {
-    multisigPlugin(id: $address){
+query MultisigMembers($address: ID!, $block: Block_height) {
+    multisigPlugin(id: $address, block: $block){
         members {
             address
         }