+++ modules/client/test/integration/addresslistVoting-client/methods.test.ts
@@ -625,13 +625,14 @@ describe("Client Address List", () => {
 
       const wallets = await client.methods.getMembers(
         ADDRESS_ONE,
+        123456,
       );
 
       expect(wallets.length).toBe(1);
       expect(wallets[0]).toBe(ADDRESS_ONE);
       expect(mockedClient.request).toHaveBeenLastCalledWith(
         QueryAddresslistVotingMembers,
-        { address: ADDRESS_ONE },
+        { address: ADDRESS_ONE, block: { number: 123456 } },
       );
     });
     it("Should fetch the given proposal", async () => {